CORE-19018 move service to visible package
