+++ components/flow/flow-service/src/main/kotlin/net/corda/flow/pipeline/MetaDataUtils.kt
@@ -1,4 +1,4 @@
-package net.corda.flow.pipeline.handlers
+package net.corda.flow.pipeline
 
 import net.corda.flow.state.impl.CheckpointMetadataKeys
 import net.corda.libs.statemanager.api.Metadata
