+++ components/flow/flow-service/src/main/kotlin/net/corda/flow/pipeline/handlers/requests/FlowFailedRequestHandler.kt
@@ -8,7 +8,7 @@ import net.corda.flow.pipeline.exceptions.FlowFatalException
 import net.corda.flow.pipeline.exceptions.FlowProcessingExceptionTypes.FLOW_FAILED
 import net.corda.flow.pipeline.factory.FlowMessageFactory
 import net.corda.flow.pipeline.factory.FlowRecordFactory
-import net.corda.flow.pipeline.handlers.addTerminationKeyToMeta
+import net.corda.flow.pipeline.addTerminationKeyToMeta
 import net.corda.flow.pipeline.handlers.requests.helper.getRecords
 import net.corda.flow.pipeline.sessions.FlowSessionManager
 import net.corda.flow.pipeline.sessions.FlowSessionStateException
